# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for restaurant hall, orders, and warehouse management",
        "title": "Restaurant Management System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/shifts/close": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Close current shift",
                "parameters": [
                    {
                        "description": "close payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.closeShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Shift"
                        }
                    }
                }
            }
        },
        "/shifts/current/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "List orders in current shift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "waiter id",
                        "name": "waiter_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.OrderWithTotal"
                            }
                        }
                    }
                }
            }
        },
        "/shifts/open": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Open a new shift",
                "parameters": [
                    {
                        "description": "shift data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.openShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Shift"
                        }
                    }
                }
            }
        },
        "/tables": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "List restaurant tables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.RestaurantTable"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Create or update restaurant table",
                "parameters": [
                    {
                        "description": "table",
                        "name": "table",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RestaurantTable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.RestaurantTable"
                        }
                    }
                }
            }
        },
        "/tables/availability": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Get table availability for interval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "from datetime (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "to datetime (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "minimum seats",
                        "name": "seats_min",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.TableAvailability"
                            }
                        }
                    }
                }
            }
        },
        "/tables/{id}": {
            "delete": {
                "tags": [
                    "tables"
                ],
                "summary": "Delete restaurant table",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "table id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.OrderWithTotal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reservation_id": {
                    "type": "integer"
                },
                "shift_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "table_id": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "waiter_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.RestaurantTable": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "seats": {
                    "type": "integer"
                },
                "table_number": {
                    "type": "integer"
                }
            }
        },
        "domain.Shift": {
            "type": "object",
            "properties": {
                "actual_revenue": {
                    "type": "number"
                },
                "closed_at": {
                    "type": "string"
                },
                "closed_by": {
                    "type": "integer"
                },
                "expected_revenue": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "opened_at": {
                    "type": "string"
                },
                "opened_by": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "domain.TableAvailability": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "is_available": {
                    "type": "boolean"
                },
                "seats": {
                    "type": "integer"
                },
                "table_id": {
                    "type": "integer"
                },
                "table_number": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.closeShiftRequest": {
            "type": "object",
            "properties": {
                "closed_by": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "handlers.openShiftRequest": {
            "type": "object",
            "properties": {
                "expected_revenue": {
                    "type": "number"
                },
                "note": {
                    "type": "string"
                },
                "opened_by": {
                    "type": "integer"
                }
            }
        },
...
    }
}